[refactor] Removed unnecessary code, deleted mock data, moved fetch functions into hooks and resolved lint errors (#54)
