+++ backend/src/routes/candidates.routes.ts
@@ -8,7 +8,7 @@ router.get("/", candidates.list);
 router.post("/", candidates.create);
 router.get("/:id", candidates.getById);
 router.get("/:id/cv", candidates.getCV);
-// router.delete("/:id", candidates.remove);
+ router.delete("/:id", candidates.remove);
 
 const upload = multer({
   storage: multer.memoryStorage(),
