+++ frontend/src/api/fetchScreenings.ts
@@ -1,21 +1,6 @@
-import { API_URL } from "./fetchCandidates";
+import type { ScreeningDetails } from "@/types/screening";
+import { API_URL } from "@/utils/variables";
 
-export type ScreeningDetails = {
-  jobPostId: number;
-  title: string;
-  screenedAt: string;
-  candidates: Array<{
-    candidateId: number;
-    candidateName: string;
-    rank: number;
-    score: number;
-    qualified: boolean;
-    qualificationsMet: string[];
-    qualificationsMissing: string[];
-    summary?: string;
-    createdAt: string;
-  }>;
-};
 
 export async function getScreeningHistory(): Promise<ScreeningDetails[]> {
   const response = await fetch(API_URL + "/api/results/history");
