+++ frontend/src/utils/variables.ts
@@ -1,0 +1,1 @@
+export const API_URL = "http://localhost:3000";
